# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/v1/watcher/checkin": {
            "post": {
                "description": "get request to perform checkin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Watcher"
                ],
                "summary": "Perform checkin via timewatch.co.il",
                "operationId": "checkin",
                "parameters": [
                    {
                        "description": "User Data",
                        "name": "user_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.UserData"
                        }
                    }
                ]
            }
        },
        "/v1/watcher/checkout": {
            "post": {
                "description": "get request to perform checkout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Watcher"
                ],
                "summary": "Perform checkout via timewatch.co.il",
                "operationId": "checkout",
                "parameters": [
                    {
                        "description": "User Data",
                        "name": "user_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.UserData"
                        }
                    }
                ]
            }
        }
    },
    "definitions": {
        "model.UserData": {
            "type": "object",
            "required": [
                "company",
                "password",
                "user"
            ],
            "properties": {
                "company": {
                    "type": "string",
                    "example": "5687"
                },
                "password": {
                    "type": "string",
                    "example": "Baba"
                },
                "user": {
                    "type": "string",
                    "example": "1234"
                }
            }
        }
    }
}